Modularised the charts
