+++ src/app/app.module.ts
@@ -2,8 +2,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { RouterModule, Routes } from '@angular/router';
+import { HttpClientModule } from '@angular/common/http';
 import { ClarityModule } from '@clr/angular';
 
 import { AppComponent } from './app.component';
@@ -12,6 +11,7 @@ import { EliteBgsModule } from './elite_bgs_api/elite-bgs-api.module';
 import { BGSBotModule } from './bgsbot/bgsbot.module';
 import { AdminModule } from './admin/admin.module';
 import { MainModule } from './main/main.module';
+import { CustomChartModule } from './charts/custom-chart.module';
 import { AppRoutingModule } from './app-routing.module';
 import { ProfileComponent } from './profile/profile.component';
 import { ProfileReportComponent } from './profile/profile-report.component';
@@ -23,6 +23,7 @@ import { GuideComponent } from './guide/guide.component';
 import { DonateComponent } from './donate/donate.component';
 import { CreditsComponent } from './credits/credits.component';
 import { PageNotFoundComponent } from './page_not_found/page-not-found.component';
+import { TickComponent } from './tick/tick.component';
 
 import { SystemsService } from './services/systems.service';
 import { FactionsService } from './services/factions.service';
@@ -45,7 +46,8 @@ import { TickService } from './services/tick.service';
         GuideComponent,
         DonateComponent,
         CreditsComponent,
-        PageNotFoundComponent
+        PageNotFoundComponent,
+        TickComponent
     ],
     imports: [
         BrowserModule,
@@ -58,10 +60,23 @@ import { TickService } from './services/tick.service';
         BGSBotModule,
         AdminModule,
         MainModule,
+        CustomChartModule,
         AppRoutingModule,
         ClarityModule
     ],
-    providers: [SystemsService, FactionsService, StationsService, AuthenticationService, ServerService, ThemeService, TryAPIService, TickService],
+    exports: [
+        CustomChartModule
+    ],
+    providers: [
+        SystemsService,
+        FactionsService,
+        StationsService,
+        AuthenticationService,
+        ServerService,
+        ThemeService,
+        TryAPIService,
+        TickService
+    ],
     bootstrap: [AppComponent]
 })
 export class AppModule { }
