+++ src/app/charts/faction-influence-chart.component.html
