+++ src/app/charts/faction-state-chart.component.html
