+++ src/app/charts/system-state-chart.component.html
