+++ src/app/charts/tick-chart.component.html
