+++ src/app/main/main-routing.module.ts
@@ -8,7 +8,6 @@ import { SystemListComponent } from './systems/system-list.component';
 import { SystemViewComponent } from './systems/system-view.component';
 import { StationListComponent } from './stations/station-list.component';
 import { StationViewComponent } from './stations/station-view.component';
-import { TickComponent } from './tick/tick.component';
 
 const mainRoutes: Routes = [
     {
@@ -42,11 +41,7 @@ const mainRoutes: Routes = [
             {
                 path: 'station/:stationid',
                 component: StationViewComponent
-            },
-            {
-                path: 'tick',
-                component: TickComponent
-            },
+            }
         ]
     }
 ];
